add uint32[] indexOf & delete Array lib helper functions; recompile contracts;

+++ typechain-types/factories/contracts/core/accounts/interface/IAccountDeployContract__factory.ts
@@ -15,9 +15,9 @@ const _abi = [
       {
         components: [
           {
-            internalType: "uint256",
+            internalType: "uint32",
             name: "id",
-            type: "uint256",
+            type: "uint32",
           },
           {
             internalType: "address",
